add more checks

+++ Scripts_Uninstallers/CortexXDR-uninstall.zsh
@@ -40,15 +40,22 @@ if [[ $password ]]; then
     else
         echo "Cortex XDR uninstallation failed"
         /usr/bin/open -a "/Applications/Cortex XDR Uninstaller.app"
+        exit 1
     fi
 else
     echo "No password provided, so cannot run the automated uninstaller. Opening the uninstaller app instead"
     /usr/bin/open -a "/Applications/Cortex XDR Uninstaller.app"
 fi
 
-# now remove the uninstaller from all locations
-rm -rf "/Applications/Cortex XDR Uninstaller.app" ||:
-rm -rf "/Library/Management/PaloAlto" ||:
+# now remove the uninstaller from all locations if Cortex has been deleted.
+if [[ ! -d "/Applications/Cortex XDR.app" ]]; then
+    rm -rf "/Applications/Cortex XDR Uninstaller.app" ||:
+    rm -rf "/Library/Management/PaloAlto" ||:
+else
+    echo "Cortex uninstallation failed"
+    /usr/bin/open -a "/Applications/Cortex XDR Uninstaller.app"
+    exit 1
+fi
 
 # Try to Forget the packages if we can find a match
 /usr/sbin/pkgutil --pkgs=com.paloaltonetworks.pkg.cortex && /usr/sbin/pkgutil --forget com.paloaltonetworks.pkg.cortex
